Migrate from AWS SDK Go v1 to v2
Migrate the codebase from AWS SDK Go v1 to v2 to address deprecation
warnings and prepare for the v1 end-of-support on July 31, 2025.

Changes:
- Update go.mod to use AWS SDK v2 dependencies:
  - github.com/aws/aws-sdk-go-v2/config
  - github.com/aws/aws-sdk-go-v2/credentials
  - github.com/aws/aws-sdk-go-v2/service/ec2
  - github.com/aws/aws-sdk-go-v2/service/elasticloadbalancing
  - github.com/aws/aws-sdk-go-v2/service/elasticloadbalancingv2

- Migrate pkg/client/client.go:
  - Replace session.Session with aws.Config
  - Update service client creation to use NewFromConfig()
  - Add context.Context parameter to all API methods
  - Migrate user agent customization from handlers to v2 middleware
  - Update IRSA authentication (still automatic in v2)
  - Update DescribeRegions API call

- Migrate pkg/controller/ec2_instance_types.go:
  - Update EC2 types to use v2 types package
  - Add context to DescribeInstanceTypes calls
  - Update architecture constants to types.ArchitectureType
  - Change VCPU and GPU fields from int64 to int32

- Migrate pkg/client/fake/fake.go:
  - Update all fake methods to accept context.Context
  - Update fake responses to use v2 types
  - Fix type conversions (int32 for GPU memory)

- Update test files:
  - Update controller_test.go to use v2 types
  - Update client_test.go to use newAWSConfig()
  - Fix architecture type tests

All unit and integration tests pass with no deprecation warnings.

Fixes #24

+++ pkg/controller/controller.go
@@ -146,9 +146,9 @@ func (r *Reconciler) reconcile(ctx context.Context, machineDeployment *clusterv1
 		machineDeployment.Annotations = make(map[string]string)
 	}
 
-	machineDeployment.Annotations[cpuKey] = strconv.FormatInt(instanceTypeInfo.VCPU, 10)
+	machineDeployment.Annotations[cpuKey] = strconv.FormatInt(int64(instanceTypeInfo.VCPU), 10)
 	machineDeployment.Annotations[memoryKey] = strconv.FormatInt(instanceTypeInfo.MemoryMb, 10)
-	machineDeployment.Annotations[gpuKey] = strconv.FormatInt(instanceTypeInfo.GPU, 10)
+	machineDeployment.Annotations[gpuKey] = strconv.FormatInt(int64(instanceTypeInfo.GPU), 10)
 
 	// Parse existing labels, update architecture, and preserve user-provided labels
 	labelsMap := make(map[string]string)
